Fix: consolidate transaction types
